examples/vit/requirements.txt
torch
torchvision
pytorch-fast-transformers
linformer
vit-pytorch
numpy
pillow
scikit-learn
tqdm
zipp